Consolidate Android components and split oversized screens
New reusable components:
- PrimaryButton: the standard brand CTA (coral, white label, medium radius,
  loading state), mirroring iOS PrimaryButtonStyle. Adopted by CartScreen's
  checkout and ProfileScreen's sign-in prompt.
- OrderCard: order summary card extracted from OrderListScreen.
- SettingsItem: settings row moved out of ProfileScreen into the component
  library.

Screen slimming (pure code moves, no visual changes):
- OrderListScreen rebuilt on LoadingStateComponent/EmptyStateComponent/
  OrderCard: 264 -> 83 lines; its hand-rolled spinner/empty-state removed;
  error state gained a working Try Again retry.
- CocktailDetailScreen: section composables (main info card, ingredients,
  chips, recommendations, reviews) moved to screens/detail/: 999 -> 257 lines.
- OfflineModeScreen: cards and dialogs moved to screens/offline/:
  521 -> 146 lines.

Cleanup:
- Deleted orphan components AuthDialog.kt, ProfileCard.kt, InfoCard.kt
  (no references anywhere).
- Removed five unused component imports from HomeScreen.
- Remaining 'Cocktail Bar' brand strings (about/copyright on both platforms)
  updated to CocktailCraft.
