+++ androidApp/src/main/java/com/cocktailcraft/android/screens/offline/OfflineModeDialogs.kt
@@ -1,0 +1,35 @@
+package com.cocktailcraft.android.screens.offline
+
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import com.cocktailcraft.android.R
+
+// Confirmation dialog shown before clearing the offline cache
+@Composable
+internal fun ClearCacheDialog(
+    onConfirm: () -> Unit,
+    onDismiss: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = { Text(stringResource(R.string.clear_offline_cache_title)) },
+        text = { Text(stringResource(R.string.clear_offline_cache_message)) },
+        confirmButton = {
+            TextButton(
+                onClick = onConfirm
+            ) {
+                Text(stringResource(R.string.clear_cache))
+            }
+        },
+        dismissButton = {
+            TextButton(
+                onClick = onDismiss
+            ) {
+                Text(stringResource(R.string.cancel))
+            }
+        }
+    )
+}
