+++ androidApp/src/main/res/values/strings.xml
@@ -213,7 +213,7 @@
     <string name="profile_about">About</string>
     <string name="profile_about_app_name">CocktailCraft</string>
     <string name="profile_version_format">Version %1$s</string>
-    <string name="profile_copyright">© 2026 Cocktail Bar. All rights reserved.</string>
+    <string name="profile_copyright">© 2026 CocktailCraft. All rights reserved.</string>
 
     <!-- Reviews & dialogs -->
     <string name="review_your_name">Your Name</string>
